requirements.txt
mysql-connector-python
PyYAML
click
rich
pytest